+++ doc/change-log.md
@@ -1,3 +1,37 @@
+# 3.2.5 - [2022-03-15]
+## Added
+- `info.gratour.common.types.MovingAverageD`
+- `info.gratour.common.types.MovingAverageL`
+
+## Improvement
+- `info.gratour.common.types.rest.Reply.ILLEGAL_STATE` added
+- `info.gratour.common.types.rest.Reply.EXECUTION_ERROR` added
+- `info.gratour.common.types.rest.Reply.INVALID_TOKEN` added
+- `info.gratour.common.types.rest.RawReply.INVALID_TOKEN` added
+- `info.gratour.common.types.rest.RawReply.ILLEGAL_STATE` added
+- `info.gratour.common.types.rest.RawReply.INVALID_CONFIG` added
+- `info.gratour.common.types.rest.RawReply.RawReply(info.gratour.common.error.ErrorWithCode)` added
+- `info.gratour.common.types.rest.RawReply.err(info.gratour.common.error.ErrorWithCode)` added
+- `info.gratour.common.error.ErrorWithCode.ILLEGAL_STATE` added
+- `info.gratour.common.error.ErrorWithCode.INVALID_CONFIG` added
+- `info.gratour.common.error.ErrorWithCode.TIMEOUT` added
+- `info.gratour.common.error.ErrorWithCode.SERVICE_UNAVAILABLE` added
+- `info.gratour.common.utils.CommonUtils.wrapIndex` added
+- `info.gratour.common.utils.NetUtils.isValidDomain` added
+- `info.gratour.common.utils.NetUtils.isValidIpOrDomain` added
+- `info.gratour.common.types.rest.Reply.isEmpty` added
+- `info.gratour.common.types.rest.Reply.nonEmpty` added
+- `info.gratour.common.types.rest.Reply.SERVICE_UNAVAILABLE` added
+- `info.gratour.common.types.validate.ValidateResultReceiver.error(java.lang.String)` added
+- `info.gratour.common.utils.DateTimeUtils.DATETIME_FORMATTER_WITH_ZONE1` added
+- `info.gratour.common.utils.DateTimeUtils.DATETIME_FORMATTER_WITH_MILLIS_ZONE1` added
+- `info.gratour.common.utils.NetUtils.HttpUtils` added
+
+## Changed
+- Dependencies:
+  - Bump `scala-library` from `2.13.6` to `2.13.8`
+  - Bump `scala-reflect` from `2.13.6` to `2.13.8`
+
 # 3.2.4 - [2022-01-02]
 ## Improvement
 - `StringUtils.strMaxLen()` added.